pyproject.toml
[project]
name = "04-file-analyzer-and-insight-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.13.0.1",
    "fitz>=0.0.1.dev2",
    "mcp[cli]>=1.19.0",
    "numpy>=2.3.4",
    "openai>=2.6.1",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "python-dotenv>=1.2.1",
    "seaborn>=0.13.2",
    "streamlit>=1.50.0",
]
